# docs/04-api/openapi-v1-core.yaml
openapi: 3.1.0
info:
  title: KWZY AI Smart Park API
  version: 1.1.0
  description: |
    AI 智慧园区平台核心主链 OpenAPI v1.1（05.1 修订）。
    统一 snake_case；统一响应 envelope；Payment=收款登记；Bill 无 OVERDUE 主状态。
    覆盖：认证、园区/单元、合同、账单、收款催缴、招商 CRM 漏斗与线索转化。
    响应约定：成功/失败均为 {code, message, data}；错误见 ErrorResponse。
    路径与字段一律 snake_case。
servers:
  - url: http://localhost:8000/api/v1
    description: Local
...
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: 线索列表
      operationId: listLeads
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: park_id
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LeadStageFilter'
        - name: keyword
          in: query
          schema: { type: string }
        - { name: owner_user_id, in: query, schema: { type: integer, format: int64 } }
        - { name: pool_status, in: query, schema: { $ref: '#/components/schemas/LeadPoolStatus' } }
        - { name: source_type, in: query, schema: { type: string } }
        - { name: created_from, in: query, schema: { type: string, format: date-time } }
        - { name: created_to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadPage'
    post:
      tags: [Leads]
      summary: 创建线索
      operationId: createLead
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreate'
      responses:
        '200':
          description: OK（统一 envelope）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
  /leads/duplicates/check:
    post:
      tags: [Leads]
      summary: 查询租户授权园区范围内的重复线索候选
      operationId: checkLeadDuplicates
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadDuplicateQuery' }
      responses:
        '200': { description: 已脱敏且按置信原因排序的候选列表 }
        '403':
          description: 无查看权限
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /crm/assignees:
    get:
      tags: [Leads]
      summary: 获取当前租户可选负责人（经理权限）
      operationId: listCrmAssignees
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 可选负责人列表 }
        '403': { description: 无线索管理权限 }
  /crm/assignment-rules:
    get:
      tags: [Leads]
      summary: 查询自动分配规则与版本状态
      operationId: listLeadAssignmentRules
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: park_id, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: 当前租户授权园区的规则列表 }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Leads]
      summary: 创建自动分配规则草稿
      operationId: createLeadAssignmentRule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentRuleCreate' } } }
      responses:
        '200': { description: 规则草稿与首个不可变版本 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/assignment-rules/preview:
    get:
      tags: [Leads]
      summary: 预演当前已发布规则的确定性归属结果
      operationId: previewLeadAssignmentRule
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: park_id, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: trigger, in: query, required: true, schema: { $ref: '#/components/schemas/AssignmentTrigger' } }
      responses:
        '200': { description: 规则版本、候选容量与选中成员证据 }
  /crm/assignment-rules/{rule_id}:
    parameters:
      - { name: rule_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询自动分配规则详情和历史版本
      operationId: getLeadAssignmentRule
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 规则详情、成员和版本链 }
  /crm/assignment-rules/{rule_id}/draft:
    parameters:
      - { name: rule_id, in: path, required: true, schema: { type: integer, format: int64 } }
    patch:
      tags: [Leads]
      summary: 更新尚未发布的规则草稿
      operationId: updateLeadAssignmentRuleDraft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentRuleDraftUpdate' } } }
      responses:
        '200': { description: 已更新草稿 }
        '409': { $ref: '#/components/responses/Conflict' }
    post:
      tags: [Leads]
      summary: 从已发布版本创建下一草稿
      operationId: createLeadAssignmentRuleDraft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExpectedLockVersion' } } }
      responses:
        '200': { description: 新的不可变版本草稿 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/assignment-rules/{rule_id}/publish:
    parameters:
      - { name: rule_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 原子发布规则并替换同园区同触发器生效版本
      operationId: publishLeadAssignmentRule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExpectedLockVersion' } } }
      responses:
        '200': { description: 已发布规则及版本证据 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/assignment-rules/{rule_id}/retire:
    parameters:
      - { name: rule_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 退役自动分配规则
      operationId: retireLeadAssignmentRule
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExpectedLockVersion' } } }
      responses:
        '200': { description: 已退役且保留历史版本 }
        '409': { $ref: '#/components/responses/Conflict' }
  /leads/{lead_id}/viewings:
    parameters:
      - { name: lead_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询线索带看计划与结果
      operationId: listLeadViewings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 带看、房源版本快照与状态 }
    post:
      tags: [Leads]
      summary: 创建同园区多单元带看计划
      operationId: createLeadViewing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ViewingCreate' } } }
      responses:
        '200': { description: 带看已排期 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/viewings/{viewing_id}:
    parameters:
      - { name: viewing_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询带看详情
      operationId: getLeadViewing
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 带看详情 }
    patch:
      tags: [Leads]
      summary: 按版本重新排期带看
      operationId: rescheduleLeadViewing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ViewingReschedule' } } }
      responses:
        '200': { description: 已重新排期 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/viewings/{viewing_id}/transition:
    parameters:
      - { name: viewing_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 确认、完成、取消或标记爽约
      operationId: transitionLeadViewing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ViewingTransition' } } }
      responses:
        '200': { description: 已转换状态；完成操作幂等写入 VISIT 活动 }
        '409': { $ref: '#/components/responses/Conflict' }
  /leads/{lead_id}/intent:
    parameters:
      - { name: lead_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询线索当前意向及权威审批状态
      operationId: getLeadIntent
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 意向快照、版本链和审批深链 }
    post:
      tags: [Leads]
      summary: 冻结首个多单元商业意向快照
      operationId: createLeadIntent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IntentCreate' } } }
      responses:
        '200': { description: 意向草稿已创建 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/intents/{intent_id}:
    parameters:
      - { name: intent_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询意向详情与不可变版本链
      operationId: getLeadIntentDetail
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 意向详情 }
  /crm/intents/{intent_id}/versions:
    parameters:
      - { name: intent_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 在草稿期追加不可变意向版本
      operationId: createLeadIntentVersion
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IntentVersionCreate' } } }
      responses:
        '200': { description: 新版本已冻结 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/intents/{intent_id}/submit:
    parameters:
      - { name: intent_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 提交统一审批并以审批中心为权威结论
      operationId: submitLeadIntent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IntentSubmit' } } }
      responses:
        '200': { description: 已创建审批申请和任务 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/channels:
    get:
      tags: [Leads]
      summary: 查询签名渠道配置（不返回明文密钥）
      operationId: listLeadChannels
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 渠道连接与本地合约验证状态 }
    post:
      tags: [Leads]
      summary: 创建默认关闭的签名渠道
      operationId: createLeadChannel
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChannelCreate' } } }
      responses:
        '200': { description: 渠道已创建；密钥仅由环境变量引用 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/channels/{channel_id}:
    parameters:
      - { name: channel_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询渠道及最近接收事件
      operationId: getLeadChannel
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 渠道配置和脱敏事件摘要 }
    patch:
      tags: [Leads]
      summary: 按版本更新、启停或轮换渠道密钥引用
      operationId: updateLeadChannel
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChannelUpdate' } } }
      responses:
        '200': { description: 渠道已更新 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/channels/{channel_id}/events:
    parameters:
      - { name: channel_id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Leads]
      summary: 查询渠道接收箱与隔离事件
      operationId: listLeadChannelEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [ACCEPTED, QUARANTINED, REJECTED] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 100 } }
      responses:
        '200': { description: 仅返回安全预览和失败码，不返回密文或明文 PII }
  /crm/channels/{channel_id}/events/{event_id}/replay:
    parameters:
      - { name: channel_id, in: path, required: true, schema: { type: integer, format: int64 } }
      - { name: event_id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Leads]
      summary: 使用当前授权配置解密并幂等重放隔离事件
      operationId: replayLeadChannelEvent
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: 重放结果和原事件证据 }
        '409': { $ref: '#/components/responses/Conflict' }
  /public/lead-channels/{public_id}/events:
    parameters:
      - { name: public_id, in: path, required: true, schema: { type: string } }
    post:
      tags: [Leads]
      summary: 接收经 HMAC v1 认证的限长渠道事件
      operationId: receiveLeadChannelEvent
      security: []
      parameters:
        - { name: X-KWZY-Timestamp, in: header, required: true, schema: { type: string } }
        - { name: X-KWZY-Event-Id, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: X-KWZY-Signature, in: header, required: true, schema: { type: string, pattern: '^v1=[0-9a-f]{64}$' } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadChannelEvent' }
      responses:
        '200': { description: 已接收或命中相同事件幂等结果 }
        '401': { description: 签名、时钟窗、重复头或长度校验失败 }
        '409': { $ref: '#/components/responses/Conflict' }
  /crm/unit-locks/sweep:
    post:
      tags: [Leads]
      summary: 清理已到期房源锁并恢复合法空置投影
      operationId: sweepCrmUnitLocks
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: park_id, in: query, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: 清理数量与 UTC 截止时间 }
        '403': { description: 无线索管理权限 }
  /crm/summary:
    get:
      tags: [Leads]
      summary: 获取与线索列表同口径的招商漏斗指标
      operationId: getCrmSummary
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/LeadStageFilter' } }
        - { name: park_id, in: query, schema: { type: integer, format: int64 } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: owner_user_id, in: query, schema: { type: integer, format: int64 } }
        - { name: pool_status, in: query, schema: { $ref: '#/components/schemas/LeadPoolStatus' } }
        - { name: source_type, in: query, schema: { type: string } }
        - { name: created_from, in: query, schema: { type: string, format: date-time } }
        - { name: created_to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: 汇总结果
          content: { application/json: { schema: { $ref: '#/components/schemas/CrmSummary' } } }
  /crm/board:
    get:
      tags: [Leads]
      summary: 获取与线索列表同口径的招商阶段看板
      operationId: getCrmBoard
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/LeadStageFilter' } }
        - { name: park_id, in: query, schema: { type: integer, format: int64 } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: owner_user_id, in: query, schema: { type: integer, format: int64 } }
        - { name: pool_status, in: query, schema: { $ref: '#/components/schemas/LeadPoolStatus' } }
        - { name: source_type, in: query, schema: { type: string } }
        - { name: created_from, in: query, schema: { type: string, format: date-time } }
        - { name: created_to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: 阶段列及线索
          content: { application/json: { schema: { $ref: '#/components/schemas/CrmBoard' } } }
...
  /leads/{lead_id}/activities:
    post:
      tags: [Leads]
      summary: 追加线索活动并可推进阶段
      operationId: addLeadActivity
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: lead_id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LeadActivityCreate' } } }
      responses:
        '200': { description: 活动已追加 }
        '409': { description: 线索版本冲突 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  schemas:
...
    LeadStageFilter:
      type: string
      enum: [NEW, CONTACTING, VISITING, QUOTING, NEGOTIATING, WON, LOST, CANCELLED, MERGED, FOLLOWING]
      description: FOLLOWING 仅为旧客户端兼容输入，服务端确定性映射为 CONTACTING；响应永不返回 FOLLOWING。
      x-legacy-value-mapping:
        FOLLOWING: CONTACTING
    LeadPoolStatus:
      type: string
      enum: [PUBLIC, PRIVATE]
...
    LeadDuplicateQuery:
      type: object
      required: [park_id, name, contact_phone]
      properties:
        park_id: { type: integer, format: int64 }
        name: { type: string, minLength: 1, maxLength: 128 }
        contact_phone: { type: string, minLength: 5, maxLength: 32 }
        source_type: { type: string, default: MANUAL }
        source_ref: { type: string, nullable: true }
...
    AssignmentTrigger:
      type: string
      enum: [MANUAL_CREATE, CHANNEL_INTAKE, RECYCLE]
...
    LeadPage:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/Lead' }
...